+++ test/unit/lib/configResponseTests.js
@@ -37,7 +37,7 @@ state.originalConfig = {
             },
             DNS: {
                 nameServers: [
-                    '8.8.8.8'
+                    '192.0.2.10'
                 ]
             },
             VLAN: {
@@ -59,7 +59,7 @@ state.originalConfig = {
             },
             DNS: {
                 nameServers: [
-                    '127.0.1.1'
+                    '192.0.2.20'
                 ]
             },
             VLAN: {
@@ -123,7 +123,7 @@ describe('configResponse', () => {
                     hostname: 'bigip1.example.com',
                     Provision: { afm: 'none', ltm: 'nominal', urldb: 'none' },
                     NTP: { timezone: 'UTC' },
-                    DNS: { nameServers: ['8.8.8.8'] },
+                    DNS: { nameServers: ['192.0.2.10'] },
                     VLAN: { foo: '1234 vlan' }
                 },
                 lastUpdate: 'last update 1234'
@@ -135,7 +135,7 @@ describe('configResponse', () => {
                     hostname: 'bigip2.example.com',
                     Provision: { afm: 'nominal', ltm: 'none', urldb: 'none' },
                     NTP: { timezone: 'America/Los_Angeles' },
-                    DNS: { nameServers: ['127.0.1.1'] },
+                    DNS: { nameServers: ['192.0.2.20'] },
                     VLAN: { foo: '5678 vlan' }
                 },
                 lastUpdate: 'last update 5678'