+++ test/unit/schema/baseSchemaTests.js
@@ -28,6 +28,7 @@ const authSchema = require('../../../src/schema/latest/auth.schema.json');
 const customFormats = require('../../../src/schema/latest/formats');
 const defSchema = require('../../../src/schema/latest/definitions.schema.json');
 const gslbSchema = require('../../../src/schema/latest/gslb.schema.json');
+const securitySchema = require('../../../src/schema/latest/security.schema.json');
 
 const ajv = new Ajv(
     {
@@ -50,6 +51,7 @@ const validate = ajv
     .addSchema(analyticsSchema)
     .addSchema(authSchema)
     .addSchema(gslbSchema)
+    .addSchema(securitySchema)
     .compile(baseSchema);
 
 /* eslint-disable quotes, quote-props */